refactor module from partition to mags
